server/requirements.txt
fastapi==0.115.12
joblib==1.4.2
numpy==2.2.6
passlib==1.7.4
pydantic==2.11.5
python-dotenv==1.1.0
python_jose==3.4.0
SQLAlchemy==2.0.41
request